+++ src/GafferSceneUIModule/GafferSceneUIModule.cpp
@@ -39,7 +39,6 @@
 #include "ContextAlgoBinding.h"
 #include "HierarchyViewBinding.h"
 #include "SceneGadgetBinding.h"
-#include "SourceSetBinding.h"
 #include "ToolBinding.h"
 #include "ViewBinding.h"
 #include "VisualiserBinding.h"
@@ -55,6 +54,5 @@ BOOST_PYTHON_MODULE( _GafferSceneUI )
 	bindHierarchyView();
 	bindSceneGadget();
 	bindContextAlgo();
-	bindSourceSet();
 
 }